DOC : responsive-ui #45 move all styled components into styles folder
흩어져있던 styled 컴포넌트를 styles 폴더로 이동함
