+++ src/components/index.ts
@@ -3,11 +3,11 @@ export { default as Modal } from './common/Modal';
 export { default as Button } from './common/Button';
 export { default as TextInput } from './common/TextInput';
 export { default as Nav } from './common/Nav';
-export { default as Member } from './common/Member';
 // Layouts
 export { default as AnonymousLayout } from './layouts/AnonymousLayout';
 export { default as AuthLayout } from './layouts/AuthLayout';
 // Member Management
+export { default as Member } from './members/Member';
 export { default as Members } from './members/Members';
 export { default as SearchID } from './members/SearchID';
 // Navbar
