+++ src/components/navbar/Groups.tsx
@@ -1,18 +1,18 @@
 import { Nav } from '..';
-import { GroupName, List } from '../../styles/componentStyle/navbar';
+import { GroupName, GroupsComponent } from '../../styles/componentStyle';
 import { useMyTeams } from '../../hooks';
 
 function Groups() {
     const { myTeams } = useMyTeams();
 
     return (
-        <List>
+        <GroupsComponent>
             {myTeams?.map((group) => (
                 <Nav key={group.teamId} url={`/teamschedule/${group.teamId}`}>
                     <GroupName>{group.name}</GroupName>
                 </Nav>
             ))}
-        </List>
+        </GroupsComponent>
     );
 }
