+++ src/styles/componentStyle/auth.ts
@@ -1,28 +1,25 @@
 import styled from '@emotion/styled/macro';
 
-const FormContainer = styled.div`
+const AuthComponent = styled.div`
     width: 300px;
     display: flex;
     flex-direction: column;
     margin: auto;
 `;
-
-const Header = styled.div`
+const AuthHeader = styled.div`
     margin: auto;
     margin-bottom: 30px;
     & h1 {
         font-size: ${({ theme }) => theme.font.size.header};
     }
 `;
-
-const FormBox = styled.form`
+const FormContainer = styled.form`
     width: 100%;
     display: flex;
     flex-direction: column;
     align-items: center;
 `;
-
-const ExplainBox = styled.div`
+const ExplainText = styled.div`
     display: flex;
     flex-direction: row;
     margin-top: 20px;
@@ -31,4 +28,4 @@ const ExplainBox = styled.div`
     font-size: small;
 `;
 
-export { FormContainer, Header, FormBox, ExplainBox };
+export { AuthComponent, FormContainer, AuthHeader, ExplainText };
